+++ pcs_test/tier0/lib/pacemaker/test_live.py
@@ -1752,16 +1752,15 @@ class HandleInstanceAttributesValidateViaPcmkTest(TestCase):
         base_cmd = ["some", "command"]
         (
             is_valid,
-            report_list,
+            reason,
         ) = lib._handle_instance_attributes_validation_via_pcmk(
             runner,
             base_cmd,
             "result/output",
             {"attr1": "val1", "attr2": "val2"},
-            not_valid_severity=Severity.info(),
         )
         self.assertTrue(is_valid)
-        self.assertEqual(report_list, [])
+        self.assertEqual(reason, "")
         runner.run.assert_called_once_with(
             base_cmd + ["--option", "attr1=val1", "--option", "attr2=val2"]
         )
@@ -1771,23 +1770,17 @@ class HandleInstanceAttributesValidateViaPcmkTest(TestCase):
         base_cmd = ["some", "command"]
         (
             is_valid,
-            report_list,
+            reason,
         ) = lib._handle_instance_attributes_validation_via_pcmk(
             runner,
             base_cmd,
             "result/output",
             {"attr1": "val1", "attr2": "val2"},
-            not_valid_severity=Severity.info(),
         )
         self.assertIsNone(is_valid)
-        assert_report_item_list_equal(
-            report_list,
-            [
-                fixture.info(
-                    report_codes.AGENT_SELF_VALIDATION_INVALID_DATA,
-                    reason="Start tag expected, '<' not found, line 1, column 1 (<string>, line 1)",
-                )
-            ],
+        self.assertEqual(
+            reason,
+            "Start tag expected, '<' not found, line 1, column 1 (<string>, line 1)",
         )
         runner.run.assert_called_once_with(
             base_cmd + ["--option", "attr1=val1", "--option", "attr2=val2"]
@@ -1806,19 +1799,15 @@ class HandleInstanceAttributesValidateViaPcmkTest(TestCase):
         base_cmd = ["some", "command"]
         (
             is_valid,
-            report_list,
+            reason,
         ) = lib._handle_instance_attributes_validation_via_pcmk(
             runner,
             base_cmd,
             "result/output",
             {"attr1": "val1", "attr2": "val2"},
-            not_valid_severity=Severity.info(),
         )
         self.assertTrue(is_valid)
-        assert_report_item_list_equal(
-            report_list,
-            [],
-        )
+        self.assertEqual(reason, "")
         runner.run.assert_called_once_with(
             base_cmd + ["--option", "attr1=val1", "--option", "attr2=val2"]
         )
@@ -1837,23 +1826,15 @@ class HandleInstanceAttributesValidateViaPcmkTest(TestCase):
         base_cmd = ["some", "command"]
         (
             is_valid,
-            report_list,
+            reason,
         ) = lib._handle_instance_attributes_validation_via_pcmk(
             runner,
             base_cmd,
             "result/output",
             {"attr1": "val1", "attr2": "val2"},
-            not_valid_severity=Severity.info(),
         )
         self.assertFalse(is_valid)
-        assert_report_item_list_equal(
-            report_list,
-            [
-                fixture.info(
-                    report_codes.AGENT_SELF_VALIDATION_RESULT, result=""
-                )
-            ],
-        )
+        self.assertEqual(reason, "")
         runner.run.assert_called_once_with(
             base_cmd + ["--option", "attr1=val1", "--option", "attr2=val2"]
         )
@@ -1881,23 +1862,17 @@ class HandleInstanceAttributesValidateViaPcmkTest(TestCase):
         base_cmd = ["some", "command"]
         (
             is_valid,
-            report_list,
+            reason,
         ) = lib._handle_instance_attributes_validation_via_pcmk(
             runner,
             base_cmd,
             "result/output",
             {"attr1": "val1", "attr2": "val2"},
-            not_valid_severity=Severity.info(),
         )
         self.assertFalse(is_valid)
-        assert_report_item_list_equal(
-            report_list,
-            [
-                fixture.info(
-                    report_codes.AGENT_SELF_VALIDATION_RESULT,
-                    result="first line\nImportant output\nand another line",
-                )
-            ],
+        self.assertEqual(
+            reason,
+            "first line\nImportant output\nand another line",
         )
         runner.run.assert_called_once_with(
             base_cmd + ["--option", "attr1=val1", "--option", "attr2=val2"]
@@ -1925,23 +1900,17 @@ class HandleInstanceAttributesValidateViaPcmkTest(TestCase):
         base_cmd = ["some", "command"]
         (
             is_valid,
-            report_list,
+            reason,
         ) = lib._handle_instance_attributes_validation_via_pcmk(
             runner,
             base_cmd,
             "result/output",
             {"attr1": "val1", "attr2": "val2"},
-            not_valid_severity=Severity.info(),
         )
         self.assertTrue(is_valid)
-        assert_report_item_list_equal(
-            report_list,
-            [
-                fixture.warn(
-                    report_codes.AGENT_SELF_VALIDATION_RESULT,
-                    result="first line\nImportant output\nand another line",
-                )
-            ],
+        self.assertEqual(
+            reason,
+            "first line\nImportant output\nand another line",
         )
         runner.run.assert_called_once_with(
             base_cmd + ["--option", "attr1=val1", "--option", "attr2=val2"]
@@ -1953,7 +1922,6 @@ class ValidateResourceInstanceAttributesViaPcmkTest(TestCase):
     def setUp(self):
         self.runner = mock.Mock()
         self.attrs = dict(attra="val1", attrb="val2")
-        self.severity = Severity.info()
         patcher = mock.patch(
             "pcs.lib.pacemaker.live._handle_instance_attributes_validation_via_pcmk"
         )
@@ -1967,7 +1935,7 @@ class ValidateResourceInstanceAttributesViaPcmkTest(TestCase):
         )
         self.assertEqual(
             lib._validate_resource_instance_attributes_via_pcmk(
-                self.runner, agent, self.attrs, self.severity
+                self.runner, agent, self.attrs
             ),
             self.ret_val,
         )
@@ -1987,7 +1955,6 @@ class ValidateResourceInstanceAttributesViaPcmkTest(TestCase):
             ],
             "./resource-agent-action/command/output",
             self.attrs,
-            self.severity,
         )
 
     def test_without_provider(self):
@@ -1996,7 +1963,7 @@ class ValidateResourceInstanceAttributesViaPcmkTest(TestCase):
         )
         self.assertEqual(
             lib._validate_resource_instance_attributes_via_pcmk(
-                self.runner, agent, self.attrs, self.severity
+                self.runner, agent, self.attrs
             ),
             self.ret_val,
         )
@@ -2014,7 +1981,6 @@ class ValidateResourceInstanceAttributesViaPcmkTest(TestCase):
             ],
             "./resource-agent-action/command/output",
             self.attrs,
-            self.severity,
         )
 
 
@@ -2024,7 +1990,6 @@ class ValidateStonithInstanceAttributesViaPcmkTest(TestCase):
     def setUp(self):
         self.runner = mock.Mock()
         self.attrs = dict(attra="val1", attrb="val2")
-        self.severity = Severity.info()
         patcher = mock.patch(
             "pcs.lib.pacemaker.live._handle_instance_attributes_validation_via_pcmk"
         )
@@ -2038,7 +2003,7 @@ class ValidateStonithInstanceAttributesViaPcmkTest(TestCase):
         )
         self.assertEqual(
             lib._validate_stonith_instance_attributes_via_pcmk(
-                self.runner, agent, self.attrs, self.severity
+                self.runner, agent, self.attrs
             ),
             self.ret_val,
         )
@@ -2054,5 +2019,4 @@ class ValidateStonithInstanceAttributesViaPcmkTest(TestCase):
             ],
             "./validate/command/output",
             self.attrs,
-            self.severity,
         )